restructured problems subpackage
