+++ docs/source/index.rst
@@ -57,7 +57,7 @@ To learn how to use ProbNum check out the `quickstart guide <introduction/quicks
    public_api/filtsmooth
    public_api/linalg
    public_api/linops
-   public_api/problem_zoo
+   public_api/problems
    public_api/quad
    public_api/random_variables
    public_api/utils
