--- docs/source/public_api/problem_zoo.rst
@@ -1,24 +1,0 @@
-probnum.problem_zoo
-========================
-
-Problem zoo for probabilistic numerical methods.
-
-Subpackage offering implementations of standard example problems or convenient
-interfaces to benchmark problems for probabilistic numerical methods. These test
-problems are meant for rapid experimentation and prototyping.
-
-.. automodapi:: probnum.problem_zoo
-    :no-heading:
-    :no-main-docstr:
-
-probnum.problem_zoo.ode
-------------------------
-
-.. automodapi:: probnum.problem_zoo.ode
-    :no-heading:
-
-probnum.problem_zoo.linalg
----------------------------
-
-.. automodapi:: probnum.problem_zoo.linalg
-    :no-heading:
