+++ docs/source/public_api/problems.rst
@@ -1,0 +1,27 @@
+probnum.problems
+================
+
+Definitions and collection of problems solved by probabilistic numerical methods.
+
+.. automodapi:: probnum.problems
+    :no-heading:
+    :no-main-docstr:
+
+probnum.problems.zoo
+********************
+
+Subpackage offering implementations of standard example problems or convenient
+interfaces to benchmark problems for probabilistic numerical methods. These test
+problems are meant for rapid experimentation and prototyping.
+
+probnum.problems.zoo.diffeq
+---------------------------
+
+.. automodapi:: probnum.problems.zoo.diffeq
+    :no-heading:
+
+probnum.problems.zoo.linalg
+---------------------------
+
+.. automodapi:: probnum.problems.zoo.linalg
+    :no-heading:
