--- src/probnum/problem_zoo/__init__.py
@@ -1,1 +1,0 @@
-"""Test problems for probabilistic linear solvers."""
