+++ src/probnum/problems/__init__.py
@@ -1,0 +1,1 @@
+"""Definitions and collection of problems solved by probabilistic linear solvers."""
