+++ src/probnum/problems/_problems.py
@@ -1,0 +1,1 @@
+"""Definitions of problems solved by probabilistic linear solvers."""
