+++ src/probnum/problems/zoo/__init__.py
@@ -1,0 +1,1 @@
+"""Collection of test problems for probabilistic linear solvers."""
