+++ tox.ini
@@ -68,7 +68,7 @@ commands =
          pylint src/probnum/linops --jobs=0
          pylint src/probnum/quad --disable="too-few-public-methods" --jobs=0
          pylint src/probnum/random_variables --disable="missing-function-docstring" --jobs=0
-         pylint src/probnum/problem_zoo --jobs=0
+         pylint src/probnum/problems --jobs=0
          pylint src/probnum/utils --disable="missing-module-docstring,missing-function-docstring" --jobs=0
          pylint benchmarks --disable="missing-function-docstring" --jobs=0  # not a work in progress, but final
          pylint tests --disable="line-too-long,duplicate-code,missing-class-docstring,unnecessary-pass,unused-variable,protected-access,attribute-defined-outside-init,no-self-use,abstract-class-instantiated,too-many-arguments,too-many-instance-attributes,too-many-locals,unused-argument,fixme,missing-module-docstring,missing-function-docstring" --jobs=0
